refactor(client): reorganise to feature-based folder structure
Move components and containers into board-editor/, game/, and lobby/
feature directories; update all import paths and routes accordingly.
Remove stale CSS from multiplayer-game that is now handled by game-board.
