+++ apps/client/src/app/board-editor/board-canvas/board-canvas.component.html
