+++ apps/client/src/app/board-editor/board-controls/board-controls.component.scss
