+++ apps/client/src/app/board-editor/board-creator/board-creator.component.scss
