+++ apps/client/src/app/board-editor/board-creator/board-creator.component.ts
@@ -1,7 +1,7 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { BoardStore, BoardService } from '@pokemon-duel/board';
-import { BoardCanvasComponent } from '../../components/board-canvas/board-canvas.component';
-import { BoardControlsComponent } from '../../components/board-controls/board-controls.component';
+import { BoardCanvasComponent } from '../board-canvas/board-canvas.component';
+import { BoardControlsComponent } from '../board-controls/board-controls.component';
 
 @Component({
   selector: 'app-board-creator',
