+++ apps/client/src/app/game/passage/passage.component.scss
