+++ apps/client/src/app/game/pokemon/pokemon.component.html
