+++ apps/client/src/app/game/pokemon/pokemon.component.scss
