+++ apps/client/src/app/game/spot/spot.component.scss
