+++ apps/client/src/app/game/spot/spot.component.spec.ts
