+++ apps/client/src/app/game/spot/spot.component.ts
