+++ apps/client/src/app/lobby/lobby/lobby.component.scss
